ai-service/requirements.txt
# AI Recommendation Service - Python Dependencies
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1

# Database
pymongo==4.10.1
motor==3.7.0

# Machine Learning Core
numpy==2.2.1
scipy==1.15.0
scikit-learn==1.6.1
pandas==2.2.3

# Recommendation specific
scikit-surprise==1.1.4
mlxtend==0.23.3

# Deep Learning (Neural Collaborative Filtering)
torch==2.5.1
--extra-index-url https://download.pytorch.org/whl/cpu

# Vector Search
faiss-cpu==1.9.0.post1

# NLP & Embeddings
sentence-transformers==3.3.1

# Job Scheduling
apscheduler==3.11.0

# Logging & Monitoring
loguru==0.7.3

# HTTP client
httpx==0.28.1

# CORS
python-multipart==0.0.20

# LLM
google-generativeai==0.8.4
